Extract TypeValidator from CodeGenerator (#63)
- Create ITypeValidatorDeps interface for dependency injection
- Extract 14+ validation methods to new TypeValidator class
- Wire up TypeValidator in CodeGenerator.generate()
- Remove ~800 lines from CodeGenerator

Validation methods extracted:
- validateIncludeNotImplementationFile (ADR-010)
- validateBitmapFieldLiteral (ADR-034)
- checkArrayBounds, validateCallbackAssignment
- checkConstAssignment (ADR-013)
- validateBareIdentifierInScope (ADR-016)
- validateNoEarlyExits, validateSwitchStatement (ADR-025)
- validateEnumExhaustiveness, validateTernaryCondition
- validateNoNestedTernary, validateDoWhileCondition
- validateShiftAmount (MISRA C:2012 Rules)
